move repositories into subfolder
