+++ src/database/repositories/replay_repository.ts
@@ -1,6 +1,6 @@
 import type { Kysely } from "kysely";
 
-import type { Database, NewReplay } from "./schema";
+import type { Database, NewReplay } from "../schema";
 
 type DB = Kysely<Database>;
