+++ src/replays/database_replay_provider/database_replay_provider.ts
@@ -2,9 +2,9 @@ import { chunk } from "@common/chunk";
 import type { FileLoadResult, FileResult, PlayerInfo, Progress, ReplayProvider } from "@replays/types";
 import type { StadiumStatsType, StatsType } from "@slippi/slippi-js";
 import { SlippiGame } from "@slippi/slippi-js";
-import * as GameRepository from "database/game_repository";
-import * as PlayerRepository from "database/player_repository";
-import * as ReplayRepository from "database/replay_repository";
+import * as GameRepository from "database/repositories/game_repository";
+import * as PlayerRepository from "database/repositories/player_repository";
+import * as ReplayRepository from "database/repositories/replay_repository";
 import type { Database, NewGame, NewPlayer, NewReplay } from "database/schema";
 import log from "electron-log";
 import * as fs from "fs-extra";
